refactor(registrar): split registration manager from proof handler
